requirements.txt
Django>=4.2,<5.0
playwright>=1.40.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
Pillow>=10.0.0
celery>=5.3.0
redis>=5.0.0
django-celery-results>=2.5.0
jsonschema>=4.20.0
openpyxl>=3.1.0
requests>=2.31.0
